optimization in C1 LIR

+++ src/hotspot/share/memory/universe.cpp
@@ -93,7 +93,7 @@
 
 pthread_mutex_t Universe::mutex_heap_event = PTHREAD_MUTEX_INITIALIZER;
 Universe::HeapEvent Universe::heap_events[1+Universe::max_heap_events] = {};
-uint32_t* Universe::heap_event_counter_ptr = (uint32_t*)&Universe::heap_events[0].heap_event_type;
+uint64_t* Universe::heap_event_counter_ptr = (uint64_t*)&Universe::heap_events[0].heap_event_type;
 bool Universe::enable_heap_event_logging = true;
 bool Universe::enable_heap_graph_verify = false && Universe::enable_heap_event_logging;
 bool Universe::heap_event_stub_in_C1_LIR = true && Universe::enable_heap_event_logging;
