+++ src/hotspot/share/memory/universe.hpp
@@ -212,7 +212,7 @@ class Universe: AllStatic {
   };
   static const int LOG_MAX_EVENT_COUNTER = 24;
   static const int max_heap_events = 1L << LOG_MAX_EVENT_COUNTER;
-  static uint32_t *heap_event_counter_ptr;
+  static uint64_t *heap_event_counter_ptr;
   static HeapEvent heap_events[1+max_heap_events];
   static pthread_mutex_t mutex_heap_event;
   static bool enable_heap_event_logging;
@@ -228,7 +228,7 @@ class Universe: AllStatic {
     // if (event.address.src == 0x0) {
     //   printf("src 0x%lx dst 0x%lx\n", event.address.src, event.address.dst);
     // }
-    uint32_t v = *Universe::heap_event_counter_ptr;
+    uint64_t v = *Universe::heap_event_counter_ptr;
     Universe::heap_events[1+v] = event;
     *Universe::heap_event_counter_ptr = v + 1;
     // if (event.heap_event_type == 0) {
